grocerycompass/apps/api/requirements.txt
# Web framework
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# Database
sqlalchemy[asyncio]==2.0.27
asyncpg==0.29.0
alembic==1.13.1
geoalchemy2==0.14.3

# Validation
pydantic==2.6.1
pydantic-settings==2.1.0

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Task queue
celery[redis]==5.3.6
redis==5.0.1

# Scraping
playwright==1.41.2
httpx==0.27.0
beautifulsoup4==4.12.3
lxml==5.1.0

# Matching & normalization
rapidfuzz==3.6.1

# Search
meilisearch==0.31.0

# Utils
python-dotenv==1.0.1
